Tests: Allow disabling bank auth and consent deletion

+++ test/OpenBanking.Library.Connector.BankTests/FunctionalSubtests/AccountAndTransaction/AccountAccessConsent/AccountAccessConsentSubtest.cs
@@ -80,17 +80,21 @@ public class AccountAccessConsentSubtest
         // Handle "normal" case
         else
         {
-            // Create fresh AccountAccessConsent
-            Guid accountAccessConsentId1 =
-                await CreateAccountAccessConsent(accountAccessConsentRequest, requestBuilder);
+            // Test creation and deletion of fresh consent
+            if (bankProfile.AccountAndTransactionApiSettings.UseAccountAccessConsentDeleteEndpointBeforeAuth)
+            {
+                // Create fresh AccountAccessConsent
+                Guid accountAccessConsentId1 =
+                    await CreateAccountAccessConsent(accountAccessConsentRequest, requestBuilder);
 
-            // Read account access consent
-            await ReadAccountAccessConsent(modifiedBy, requestBuilder, accountAccessConsentId1);
+                // Read account access consent
+                await ReadAccountAccessConsent(modifiedBy, requestBuilder, accountAccessConsentId1);
 
-            // Delete AccountAccessConsent (includes external API delete)
-            await DeleteAccountAccessConsent(modifiedBy, requestBuilder, accountAccessConsentId1, true);
+                // Delete AccountAccessConsent (includes external API delete)
+                await DeleteAccountAccessConsent(modifiedBy, requestBuilder, accountAccessConsentId1, true);
+            }
 
-            // Use existing AccountAccessConsent for further testing (to avoid creating orphan object at bank if test terminates)
+            // Perform further testing with existing AccountAccessConsent (to avoid creating orphan object at bank if test terminates)
             if (testData2.AccountAccessConsentExternalApiId is not null)
             {
                 // Create AccountAccessConsent using existing external API consent
@@ -103,7 +107,8 @@ public class AccountAccessConsentSubtest
                 await ReadAccountAccessConsent(modifiedBy, requestBuilder, accountAccessConsentId2);
 
                 // Consent authorisation
-                if (consentAuth is not null)
+                if (consentAuth is not null &&
+                    testData2.AuthDisable is not true)
                 {
                     // Authorise consent in UI via Playwright
                     async Task<bool> AuthIsComplete()
